Use Next/Dynamic to fix WordCloud Reference Error
According to this Stack Overflow Post (see below), Modules that only work in the browser - i.e. depend on things like document or window - will not work if Next tries to perform it using Server side Rendering. So as a result I implemented dynamic load of the Word Cloud library instead and that seemed to work.

Link to Stack Overflow Post:
https://stackoverflow.com/questions/60629258/next-js-document-is-not-defined
(Taken 07/09/2023 at 16:29 GMT)